[#250] refactor: category 패키지 구조 이동
