+++ linkmind/src/main/java/com/app/toaster/category/controller/request/ChangeCateoryTitleDto.java
@@ -1,4 +1,4 @@
-package com.app.toaster.controller.request.category;
+package com.app.toaster.category.controller.request;
 
 import com.app.toaster.controller.valid.Severity;
 import com.app.toaster.controller.valid.TitleValid;
