+++ linkmind/src/main/java/com/app/toaster/category/controller/response/DuplicatedResponse.java
@@ -1,4 +1,4 @@
-package com.app.toaster.controller.response.category;
+package com.app.toaster.category.controller.response;
 
 public record DuplicatedResponse(Boolean isDupicated) {
 	public static DuplicatedResponse of(Boolean isDupicated){
