+++ linkmind/src/main/java/com/app/toaster/category/service/CategoryService.java
@@ -1,13 +1,15 @@
-package com.app.toaster.service.category;
+package com.app.toaster.category.service;
 
-import com.app.toaster.controller.request.category.*;
-import com.app.toaster.controller.response.category.CategoriesResponse;
-import com.app.toaster.controller.response.category.DuplicatedResponse;
+import com.app.toaster.category.controller.request.ChangeCateoryPriorityDto;
+import com.app.toaster.category.controller.request.ChangeCateoryTitleDto;
+import com.app.toaster.category.controller.request.CreateCategoryDto;
+import com.app.toaster.category.controller.response.CategoriesResponse;
+import com.app.toaster.category.controller.response.DuplicatedResponse;
 import com.app.toaster.toast.controller.response.ToastDto;
 import com.app.toaster.toast.controller.response.ToastFilter;
-import com.app.toaster.controller.response.category.CategoryResponse;
-import com.app.toaster.controller.response.category.GetCategoryResponseDto;
-import com.app.toaster.domain.Category;
+import com.app.toaster.category.controller.response.CategoryResponse;
+import com.app.toaster.category.controller.response.GetCategoryResponseDto;
+import com.app.toaster.category.domain.Category;
 import com.app.toaster.timer.domain.Reminder;
 import com.app.toaster.toast.domain.Toast;
 import com.app.toaster.user.domain.User;
@@ -15,7 +17,7 @@ import com.app.toaster.exception.Error;
 import com.app.toaster.exception.model.BadRequestException;
 import com.app.toaster.exception.model.CustomException;
 import com.app.toaster.exception.model.NotFoundException;
-import com.app.toaster.infrastructure.CategoryRepository;
+import com.app.toaster.category.infrastructure.CategoryRepository;
 import com.app.toaster.timer.infrastructure.TimerRepository;
 import com.app.toaster.toast.infrastructure.ToastRepository;
 import com.app.toaster.user.infrastructure.UserRepository;
