+++ linkmind/src/main/java/com/app/toaster/common/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.app.toaster.domain;
+package com.app.toaster.common.domain;
 
 
 import jakarta.persistence.EntityListeners;
