+++ linkmind/src/main/java/com/app/toaster/common/domain/IntegerListConverter.java
@@ -1,4 +1,4 @@
-package com.app.toaster.domain;
+package com.app.toaster.common.domain;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
