+++ linkmind/src/main/java/com/app/toaster/controller/response/search/SearchCategoryResult.java
@@ -2,8 +2,6 @@ package com.app.toaster.controller.response.search;
 
 import java.util.List;
 
-import com.app.toaster.domain.Category;
-
 public record SearchCategoryResult(List<CategoryResult> categories) {
 	public static SearchCategoryResult of(List<CategoryResult> categories){
 		return new SearchCategoryResult(categories);
