--- linkmind/src/main/java/com/app/toaster/service/category/CategoryQueryService.java
@@ -1,12 +1,0 @@
-package com.app.toaster.service.category;
-
-import org.springframework.transaction.annotation.Transactional;
-
-import lombok.RequiredArgsConstructor;
-
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class CategoryQueryService {
-
-	// R
-}
