+++ linkmind/src/main/java/com/app/toaster/timer/service/TimerService.java
@@ -7,7 +7,7 @@ import com.app.toaster.timer.controller.response.CompletedTimerDto;
 import com.app.toaster.timer.controller.response.GetTimerPageResponseDto;
 import com.app.toaster.timer.controller.response.GetTimerResponseDto;
 import com.app.toaster.timer.controller.response.WaitingTimerDto;
-import com.app.toaster.domain.Category;
+import com.app.toaster.category.domain.Category;
 import com.app.toaster.timer.domain.Reminder;
 import com.app.toaster.user.domain.User;
 import com.app.toaster.exception.Error;
@@ -15,7 +15,7 @@ import com.app.toaster.exception.model.CustomException;
 import com.app.toaster.exception.model.ForbiddenException;
 import com.app.toaster.exception.model.NotFoundException;
 import com.app.toaster.exception.model.UnauthorizedException;
-import com.app.toaster.infrastructure.CategoryRepository;
+import com.app.toaster.category.infrastructure.CategoryRepository;
 import com.app.toaster.timer.infrastructure.TimerRepository;
 import com.app.toaster.user.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
