+++ linkmind/src/main/java/com/app/toaster/toast/domain/Toast.java
@@ -2,7 +2,7 @@ package com.app.toaster.toast.domain;
 
 import java.time.LocalDateTime;
 
-import com.app.toaster.domain.Category;
+import com.app.toaster.category.domain.Category;
 import com.app.toaster.user.domain.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
