+++ linkmind/src/main/java/com/app/toaster/toast/service/ToastService.java
@@ -16,7 +16,7 @@ import com.app.toaster.toast.controller.request.UpdateToastDto;
 import com.app.toaster.parse.controller.response.OgResponse;
 import com.app.toaster.toast.controller.response.IsReadResponse;
 import com.app.toaster.toast.controller.response.ModifiedTitle;
-import com.app.toaster.domain.Category;
+import com.app.toaster.category.domain.Category;
 import com.app.toaster.toast.domain.Toast;
 import com.app.toaster.user.domain.User;
 import com.app.toaster.exception.Error;
@@ -25,7 +25,7 @@ import com.app.toaster.exception.model.CustomException;
 import com.app.toaster.exception.model.ForbiddenException;
 import com.app.toaster.exception.model.NotFoundException;
 import com.app.toaster.external.client.aws.S3Service;
-import com.app.toaster.infrastructure.CategoryRepository;
+import com.app.toaster.category.infrastructure.CategoryRepository;
 import com.app.toaster.toast.infrastructure.ToastRepository;
 import com.app.toaster.user.infrastructure.UserRepository;
 import com.app.toaster.parse.service.ParsingService;
