+++ linkmind/src/test/java/com/app/toaster/infrastructure/querydsl/CustomToastRepositoryTest.java
@@ -23,7 +23,7 @@
 // import com.app.toaster.config.JpaQueryFactoryConfig;
 // import com.app.toaster.toast.domain.Toast;
 // import com.app.toaster.fixture.Fixture;
-// import com.app.toaster.infrastructure.CategoryRepository;
+// import com.app.toaster.category.infrastructure.CategoryRepository;
 // import com.app.toaster.toast.infrastructure.ToastRepository;
 // import com.app.toaster.user.infrastructure.UserRepository;
 // import com.querydsl.core.types.dsl.BooleanExpression;
